Repackaged Dave Bryson's erlang_websocket
Code has been refactored into the form of a Sinan project that can be
installed with the Faxien package manager.  In addition to simplifying
installation, this eliminates the need for the following:

  *  An embedded copy of Mochiweb.

  *  Makefiles.

The project has also been renamed to mochiweb_websocket as I feel that
better describes what it does and what it depends on.

Examples are currently excluded pending more refactoring.